% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_date}
\alias{convert_date}
\title{convert_date}
\usage{
convert_date(created_at)
}
\arguments{
\item{created_at}{the date string to convert}
}
\value{
Return the date in a datetime format
}
\description{
Convert messy dates as they come from the Twitter API in text format
}
