% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnToAccounts.R
\name{turnToAccount}
\alias{turnToAccount}
\title{turnToAccount}
\usage{
turnToAccount(df)
}
\arguments{
\item{df}{The target table}
}
\value{
A user dataset
}
\description{
Turn the 88 columns returned by rtweet into a user account dataset
}
