% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnToPost.R
\name{turnToPost}
\alias{turnToPost}
\title{turnToPost}
\usage{
turnToPost(df)
}
\arguments{
\item{df}{The target table}
}
\value{
A posts dataset
}
\description{
Turn the 88 columns returned by rtweet into a posts dataset
}
